# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Careera API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from "./event";

export interface Registration {
  id: number;
  eventId: number;
  userId: number;
  event?: Event;
  registeredAt: Date;
}